purge snapd for 20.04

+++ tests/pre-deploy.sh
@@ -24,7 +24,10 @@ lxc profile set default raw.lxc lxc.aa_profile=unconfined
 lxc profile set default security.privileged true
 lxc profile show default
 lxc launch ubuntu:${UBUNTU_VERSION} algo
-lxc exec algo -- apt remove snapd --purge -y || true
+
+if [[ ${UBUNTU_VERSION} == "20.04" ]]; then
+  lxc exec algo -- apt remove snapd --purge -y || true
+fi
 
 ip addr
